Add files via upload
fixed add_concept (it adds children to parents only if concept is not word)
added two fields to Concept object: wordlink (link to WL) and kblink (link to KB)
modified add_word and add_concept so that if concept is word then parents are empty, kblink and wordlink are filled.
added a unit test in module testing, utest_read_concept. This has 7 test cases for read_concept. 
There is a bug in read_concept, some test cases dont work, see comments in code.