branching and general versus specific reasoning
Paragraph borders detected and stored for all branches.  .g generality attribute added to concepts. g=0 means specific concept (works for words only). If a specific concept has no mapping rule, it will be automatically mapped to the first occurence in the paragraph. If a question has a word to be mapped, that is inserted before the question in a D(x) concept to force mapping.